test: refactor splitted tests into one test route
